Mark unintrospectable methods as (skip)
xkl_default_log_appender() uses varargs. If we need this in GI bindings, it
needs a xkl_default_log_appender_v() method which takes an array instead of
varargs.

xkl_set_log_appender() gets a callback which is neither (call) or (async)
scope. If this is wanted in GI bindings, it needs a new method which
additinally takes a GDestroyNotify argument.